line between closest points on cube and cylinder is drawn correctly

+++ src/ui/shape_info.c
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/02 17:32:41 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/01/29 18:44:19 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/01/30 00:36:28 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -146,7 +146,7 @@ void	dda(t_scene *scene, float x1, float x2, float y1, float y2, int color)
 	dy /= steps;
 	while (i <= steps)
 	{
-		dst = scene->disp->disp_addr + (int)(y1 * scene->settings.disp_w + x1) * scene->disp->bpp;
+		dst = scene->disp->disp_addr + (int)((int)y1 * scene->settings.disp_w + (int)x1) * scene->disp->bpp;
 		*(unsigned int*)dst = color;
 		y1 += dy;
 		x1 += dx;
